bd_website_git/requirements.txt
Django>=4.1,<4.2
gunicorn>=20.1.0,<20.2
Pillow>=9.2.0,<9.3.0
django-storages>=1.13.1,<1.14.0
django-storages[google]
django-qr-code>=3.1.1,<3.2.0
django-crispy-forms>=1.14.0,<1.15.0
psycopg2>=2.9.3,<2.10.0
stripe>=4.0.2
mysqlclient>=2.1.1,<2.2.0